implementing logic needed for tests

+++ src/configurations.py
@@ -7,11 +7,11 @@ from ideasList import IdeasList
 
 
 @dataclass
-class Saved:
+class Paths:
     RESOURCES = Path(__file__).parent / 'resources'
     DATA_DIR = RESOURCES / 'data'
-    IDEAS_NAME = 'ideas.pickle'
-    CATEGORIES_NAME = 'categories.pickle'
+    IDEAS_NAME: str = 'ideas.pickle'
+    CATEGORIES_NAME: str = 'categories.pickle'
     IDEAS_PATH = DATA_DIR / IDEAS_NAME
     CATEGORIES_PATH = DATA_DIR / CATEGORIES_NAME
 
@@ -35,9 +35,9 @@ def save_list(list_to_save: Any, path: Path):
 
 
 def get_default_list(path: Path):
-    if Saved.IDEAS_NAME in path.name:
+    if Paths.IDEAS_NAME in path.name:
         return IdeasList()
-    elif Saved.CATEGORIES_NAME in path.name:
+    elif Paths.CATEGORIES_NAME in path.name:
         return []
     return []
 
@@ -47,11 +47,12 @@ def _save_default_list(path: Path, test_mode=False):
 
 
 def _create_directories(test_mode=False):
-    dirs = (Saved.RESOURCES, Saved.DATA_DIR) if not test_mode else (Saved.RESOURCES, Saved.TESTS_DIR)
+    dirs = (Paths.RESOURCES, Paths.DATA_DIR) if not test_mode else (Paths.RESOURCES, Paths.TESTS_DIR)
     for path in dirs:
         if not path.exists():
             path.mkdir()
 
 
 def remove_file(path: Path):
-    path.unlink()
+    if path.exists():
+        path.unlink()
